requirements/python-tools.txt
pyelftools>=0.30,<1
